Revert "Refactor: Change module name to com.lambda.common"
This reverts commit af8c6a15866f85abcbd5f629d8eae2eae1a9cc18.
