+++ common/src/main/java/com/lambda/mixin/KeyboardMixin.java
@@ -1,0 +1,19 @@
+package com.lambda.mixin;
+
+import com.lambda.event.EventFlow;
+import com.lambda.event.events.KeyPressEvent;
+import net.minecraft.client.Keyboard;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(Keyboard.class)
+public class KeyboardMixin {
+    @Inject(method = "onKey", at = @At("HEAD"))
+    void onKey(long window, int key, int scancode, int action, int modifiers, CallbackInfo ci) {
+        if (key <= 0) return;
+        if (action != 1) return;
+        EventFlow.post(new KeyPressEvent(key));
+    }
+}
