+++ common/src/main/java/com/lambda/mixin/MinecraftClientMixin.java
@@ -1,0 +1,22 @@
+package com.lambda.mixin;
+
+import com.lambda.event.EventFlow;
+import com.lambda.event.events.TickEvent;
+import net.minecraft.client.MinecraftClient;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(MinecraftClient.class)
+public class MinecraftClientMixin {
+    @Inject(method = "tick", at = @At("HEAD"))
+    void onTickPre(CallbackInfo ci) {
+        EventFlow.post(new TickEvent.Pre());
+    }
+
+    @Inject(method = "tick", at = @At("RETURN"))
+    void onTickPost(CallbackInfo ci) {
+        EventFlow.post(new TickEvent.Post());
+    }
+}
