+++ common/src/main/kotlin/com/lambda/LoaderInfo.kt
@@ -1,0 +1,11 @@
+package com.lambda
+
+import dev.architectury.injectables.annotations.ExpectPlatform
+import org.jetbrains.annotations.Contract
+
+object LoaderInfo {
+    @Contract(pure = true)
+    @ExpectPlatform
+    @JvmStatic
+    fun getVersion(): String = "DEV"
+}
