--- common/src/main/kotlin/com/lambda/common/LoaderInfo.kt
@@ -1,11 +1,0 @@
-package com.lambda.common
-
-import dev.architectury.injectables.annotations.ExpectPlatform
-import org.jetbrains.annotations.Contract
-
-object LoaderInfo {
-    @Contract(pure = true)
-    @ExpectPlatform
-    @JvmStatic
-    fun getVersion(): String = "DEV"
-}
