--- common/src/main/kotlin/com/lambda/common/event/CallbackEvent.kt
@@ -1,3 +1,0 @@
-package com.lambda.common.event
-
-interface CallbackEvent : Event
