--- common/src/main/kotlin/com/lambda/common/event/Event.kt
@@ -1,3 +1,0 @@
-package com.lambda.common.event
-
-interface Event
