--- common/src/main/kotlin/com/lambda/common/event/EventFlow.kt
@@ -1,90 +1,0 @@
-package com.lambda.common.event
-
-import com.lambda.common.event.listener.Listener
-import com.lambda.common.threading.runConcurrent
-import kotlinx.coroutines.*
-import kotlinx.coroutines.channels.BufferOverflow
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.filter
-import kotlinx.coroutines.flow.filterNot
-
-object EventFlow {
-    /**
-     * [lambdaScope] is a [CoroutineScope] which is used to launch coroutines.
-     *
-     * It is defined with [Dispatchers.Default] which is optimized for CPU-intensive work
-     * such as sorting large lists, doing complex calculations and similar.
-     *
-     * The [SupervisorJob] is used so that failure or cancellation of one child does not
-     * lead to the failure or cancellation of the parent or its other children, which is
-     * useful when you have multiple independent [Job]s running in parallel.
-     */
-    val lambdaScope = CoroutineScope(Dispatchers.Default + SupervisorJob())
-    private val concurrentFlow = MutableSharedFlow<Event>(
-        extraBufferCapacity = 1000,
-        onBufferOverflow = BufferOverflow.DROP_OLDEST
-    )
-
-    val syncListeners = Subscriber()
-    val concurrentListeners = Subscriber()
-
-    init {
-        // parallel event execution on dedicated threads
-        runConcurrent {
-            concurrentFlow
-                // early filter to avoid an unnecessary collection
-                .filter { concurrentListeners.containsKey(it::class) }
-                .filterNot { it is ICancellable && it.isCanceled() }
-                .collect { event ->
-                    event.executeListenerConcurrently()
-                }
-        }
-    }
-
-    /**
-     * Posts an [Event] to the event flow [concurrentFlow] and the synchronous [Listener]s.
-     *
-     * This function first notifies all asynchronous [Listener]s by emitting the event to the [concurrentFlow].
-     * Each asynchronous [Listener] will execute its [Listener] function on a new coroutine.
-     *
-     * After notifying asynchronous [Listener]s, it executes the [Listener] functions of all synchronous [Listener]s.
-     * An instant callback ([CallbackEvent]) can only be achieved by synchronous listening objects
-     * as the concurrent listener will be executed "later".
-     *
-     * @param event The [Event] to be posted to the event flow.
-     */
-    @JvmStatic
-    fun post(event: Event) {
-        concurrentFlow.tryEmit(event)
-        event.executeListenerSynchronous()
-    }
-
-    /**
-     * Unsubscribes from both synchronous and concurrent event flows for a specific [Event] type [T].
-     *
-     * This function removes the listeners associated with the specified event type from both synchronous and concurrent event flows.
-     * After this function is called, the listeners of the specified event type will no longer be triggered when the event is dispatched.
-     *
-     * @param T The type of the event to unsubscribe from. This should be a subclass of Event.
-     */
-    inline fun <reified T : Event> unsubscribe() {
-        syncListeners.remove(T::class)
-        concurrentListeners.remove(T::class)
-    }
-
-    private fun Event.executeListenerSynchronous() {
-        syncListeners[this::class]?.forEach { listener ->
-            if (this is ICancellable && this.isCanceled()) return
-            listener.execute(this@executeListenerSynchronous)
-        }
-    }
-
-    private fun Event.executeListenerConcurrently() {
-        concurrentListeners[this::class]?.forEach { listener ->
-            if (this is ICancellable && this.isCanceled()) return
-            runConcurrent {
-                listener.execute(this@executeListenerConcurrently)
-            }
-        }
-    }
-}
