--- common/src/main/kotlin/com/lambda/common/event/events/KeyPressEvent.kt
@@ -1,6 +1,0 @@
-package com.lambda.common.event.events
-
-import com.lambda.common.event.Cancellable
-import com.lambda.common.event.ICancellable
-
-class KeyPressEvent(val key: Int): ICancellable by Cancellable()
