--- common/src/main/kotlin/com/lambda/common/event/events/TickEvent.kt
@@ -1,8 +1,0 @@
-package com.lambda.common.event.events
-
-import com.lambda.common.event.Event
-
-abstract class TickEvent : Event {
-    class Pre : TickEvent()
-    class Post : TickEvent()
-}
