+++ common/src/main/kotlin/com/lambda/event/Event.kt
@@ -1,0 +1,3 @@
+package com.lambda.event
+
+interface Event
