+++ common/src/main/kotlin/com/lambda/event/events/KeyPressEvent.kt
@@ -1,0 +1,6 @@
+package com.lambda.event.events
+
+import com.lambda.event.Cancellable
+import com.lambda.event.ICancellable
+
+class KeyPressEvent(val key: Int): ICancellable by Cancellable()
