+++ common/src/main/kotlin/com/lambda/event/events/TickEvent.kt
@@ -1,0 +1,8 @@
+package com.lambda.event.events
+
+import com.lambda.event.Event
+
+abstract class TickEvent : Event {
+    class Pre : TickEvent()
+    class Post : TickEvent()
+}
