+++ fabric/src/main/kotlin/com/lambda/fabric/LambdaFabric.kt
@@ -1,8 +1,8 @@
 package com.lambda.fabric
 
 import net.fabricmc.api.ClientModInitializer
-import com.lambda.common.Lambda
-import com.lambda.common.Lambda.LOG
+import com.lambda.Lambda
+import com.lambda.Lambda.LOG
 
 object LambdaFabric : ClientModInitializer {
     override fun onInitializeClient() {
