+++ neoforge/src/main/kotlin/com/lambda/neoforge/LambdaNeoForge.kt
@@ -1,8 +1,8 @@
 package com.lambda.neoforge
 
 import net.neoforged.fml.common.Mod
-import com.lambda.common.Lambda
-import com.lambda.common.Lambda.LOG
+import com.lambda.Lambda
+import com.lambda.Lambda.LOG
 
 @Mod(Lambda.MOD_ID)
 object LambdaNeoForge {
